props_pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prop-tracker"
version = "0.1.0"
description = "Local pipeline for sportsbook prop odds tracking and evaluation"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "pandas>=2.0.0",
  "python-dotenv>=1.0.0",
]

[project.scripts]
prop-tracker = "prop_tracker.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
